# api-docs/openapi.yaml
openapi: 3.1.0
info:
  title: DBackup API
  version: 2.0.0
  description: |
    REST API for DBackup - a self-hosted database backup automation platform with encryption, compression, and smart retention.

    ## Authentication

    DBackup supports two authentication methods:

    - **Session Authentication** - Used automatically when logged in via the web UI. Session cookies are sent with each request.
    - **API Key Authentication** - For scripts, CI/CD pipelines, and external integrations. Create an API key under **Access Management → API Keys**.

    > API keys do not inherit SuperAdmin privileges. Only explicitly assigned permissions are available.

    ## Rate Limits

    | Request Type | Default Limit |
    |---|---|
    | Authentication (login, etc.) | 5/min per IP |
    | GET requests | 100/min per IP |
    | POST / PUT / DELETE | 20/min per IP |

    Rate limits are configurable in Settings → Rate Limits.
  license:
    name: GPL-3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
...
servers:
  - url: "{protocol}://{host}/api"
    description: DBackup Instance
    variables:
      protocol:
        default: https
        enum: [http, https]
      host:
        default: your-dbackup-instance.com
security:
  - BearerAuth: []
  - SessionAuth: []
...
paths:
...
  /executions/{id}/cancel:
    post:
      tags: [Executions]
      summary: Cancel Execution
      description: |
        Cancel a pending or running execution.

        - A `Pending` execution is cancelled immediately.
        - A `Running` execution receives an abort signal. If the process is not
          running in this server instance the status is forced to `Cancelled` in
          the database.

        **Required permission:** `jobs:execute`
      operationId: cancelExecution
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Execution cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Execution is already completed and cannot be cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  # ─── History ───────────────────────────────────────────────────
...
  /adapters/{id}/databases:
    get:
      tags: [Adapters]
      summary: List Databases
      description: Fetch available database names from a saved database source.
      operationId: listAdapterDatabases
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Database list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  databases:
                    type: array
                    items:
                      type: string
                    example: ["mydb", "app_production", "analytics"]
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /adapters/test-connection:
    post:
      tags: [Connection Testing]
      summary: Test Connection
      description: Tests connectivity for a given adapter configuration. If `configId` is provided and the test succeeds, the adapter metadata is updated.
      operationId: testConnection
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [adapterId, config]
              properties:
                adapterId:
                  type: string
                  description: Adapter type identifier
                  example: mysql
                config:
                  type: object
                  description: Adapter-specific connection configuration
                configId:
                  type: string
                  description: Existing adapter ID (updates metadata on success)
                primaryCredentialId:
                  type: string
                  description: Credential profile ID to overlay on the primary connection config (username, password, access key, token, etc.)
                sshCredentialId:
                  type: string
                  description: Credential profile ID to overlay on the SSH tunnel config (SSH key or password)
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "200":
          description: Connection result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                    example: Connection successful
                  version:
                    type: string
                    example: "8.0.35"
  /adapters/access-check:
    post:
      tags: [Connection Testing]
      summary: Access Check (List Databases)
      description: Tests connection and discovers available databases.
      operationId: accessCheck
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [adapterId, config]
              properties:
                adapterId:
                  type: string
                  example: postgres
                config:
                  type: object
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "200":
          description: Databases found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  databases:
                    type: array
                    items:
                      type: string
                    example: ["mydb", "app_production"]
  /adapters/database-stats:
    post:
      tags: [Connection Testing]
      summary: Database Statistics
      description: |
        Fetches database size and table count information. Provide either `sourceId` (existing adapter) or `adapterId` + `config` (ad-hoc connection).
      operationId: getDatabaseStats
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                adapterId:
                  type: string
                  description: Required if no `sourceId`
                config:
                  type: object
                  description: Required if no `sourceId`
                sourceId:
                  type: string
                  description: Existing source adapter ID (alternative)
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "200":
          description: Database statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  databases:
                    type: array
                    items:
                      $ref: "#/components/schemas/DatabaseInfo"
                  serverVersion:
                    type: string
                    example: "8.0.35"
                  serverEdition:
                    type: string
                    example: "MySQL Community Server"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        API Key authentication. Create an API key under **Access Management → API Keys**.

        Format: `dbackup_xxxxxxxx`
    SessionAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
      description: Session cookie (browser login)
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Resource ID
  responses:
    Unauthorized:
      description: Missing, invalid, disabled, or expired credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Valid credentials but insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Human-readable error message"
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
...
    DatabaseInfo:
      type: object
      properties:
        name:
          type: string
          example: mydb
        sizeBytes:
          type: integer
          example: 1073741824
        tableCount:
          type: integer
          example: 42
...