# src/app/api/adapters/test-connection/route.ts
import { NextRequest, NextResponse } from "next/server";
import { registry } from "@/lib/core/registry";
import { registerAdapters } from "@/lib/adapters";
import { overlayCredentialsOnConfig } from "@/lib/adapters/config-resolver";
import { headers } from "next/headers";
import prisma from "@/lib/prisma";
import { getAuthContext, checkPermissionWithContext } from "@/lib/auth/access-control";
import { PERMISSIONS, Permission } from "@/lib/auth/permissions";
import { logger } from "@/lib/logging/logger";
import { wrapError } from "@/lib/logging/errors";

const log = logger.child({ route: "adapters/test-connection" });

// Ensure adapters are registered
registerAdapters();

// Helper to determine permission based on adapter type
function getPermissionForAdapter(adapterId: string): Permission | null {
    if (/mysql|postgres|mongo|mssql|sqlite/i.test(adapterId)) {
        return PERMISSIONS.SOURCES.READ;
    } else if (/local-filesystem|s3|sftp|smb|ftp|webdav|rsync|google-drive|dropbox|onedrive/i.test(adapterId)) {
        return PERMISSIONS.DESTINATIONS.READ;
    } else if (/discord|email|smtp|slack/i.test(adapterId)) {
        return PERMISSIONS.NOTIFICATIONS.READ;
    }
    return null;
}

export async function POST(req: NextRequest) {
    const ctx = await getAuthContext(await headers());

    if (!ctx) {
        return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
    }

    try {
        const body = await req.json();
        const { adapterId, config, configId, primaryCredentialId, sshCredentialId } = body;

        // RBAC: Check permission based on adapter type
        const requiredPermission = getPermissionForAdapter(adapterId || '');
        if (requiredPermission) {
            checkPermissionWithContext(ctx, requiredPermission);
        }

        if (!adapterId || !config) {
            return NextResponse.json({ success: false, message: "Missing required fields" }, { status: 400 });
        }

        const adapter = registry.get(adapterId);

        if (!adapter) {
            return NextResponse.json({ success: false, message: "Adapter not found" }, { status: 404 });
        }

        if (!adapter.test) {
            return NextResponse.json({ success: false, message: "This adapter does not support connection testing." });
        }

        // Overlay credential profiles onto the (plaintext) config so that the
        // adapter receives a fully merged, ready-to-use config. Required when
        // the user has assigned credential profiles instead of inline secrets.
        const mergedConfig = await overlayCredentialsOnConfig(
            adapterId,
            { ...config },
            primaryCredentialId ?? null,
            sshCredentialId ?? null
        );

        const result = await adapter.test(mergedConfig);

        // If test successful and we have a configId (editing existing config), update metadata
        if (result.success && result.version && configId) {
            try {
                const existingConfig = await prisma.adapterConfig.findUnique({
                    where: { id: configId },
                    select: { metadata: true }
                });

                const currentMeta = existingConfig?.metadata ? JSON.parse(existingConfig.metadata) : {};
                const newMeta = {
                    ...currentMeta,
                    engineVersion: result.version,
                    lastCheck: new Date().toISOString(),
                    status: 'Online'
                };

                await prisma.adapterConfig.update({
                    where: { id: configId },
                    data: { metadata: JSON.stringify(newMeta) }
                });
            } catch (metaError: unknown) {
                log.error("Failed to update metadata", { configId }, wrapError(metaError));
                // Don't fail the entire request if metadata update fails
            }
        }

        return NextResponse.json(result);

    } catch (error: unknown) {
        const message = error instanceof Error ? error.message : "Unknown error";
        return NextResponse.json({ success: false, message }, { status: 500 });
    }
}